+++ Demo/DemoTests/MoyaProviderSpec.swift
@@ -229,27 +229,6 @@ class MoyaProviderSpec: QuickSpec {
                 expect(receivedError?.localizedDescription) == "Houston, we have a problem"
             }
         }
-
-        describe("with lazy data") {
-            var provider: MoyaProvider<GitHub>!
-            beforeEach {
-                provider = MoyaProvider<GitHub>(endpointClosure: lazyEndpointClosure, stubClosure: MoyaProvider.ImmediatelyStub)
-            }
-
-            it("returns stubbed data for zen request") {
-                var message: String?
-
-                let target: GitHub = .Zen
-                provider.request(target) { (data, statusCode, response, error) in
-                    if let data = data {
-                        message = NSString(data: data, encoding: NSUTF8StringEncoding) as? String
-                    }
-                }
-
-                let sampleData = target.sampleData as NSData
-                expect(message).to(equal(NSString(data: sampleData, encoding: NSUTF8StringEncoding)))
-            }
-        }
     }
 }
 
@@ -294,13 +273,9 @@ private func url(route: MoyaTarget) -> String {
     return route.baseURL.URLByAppendingPathComponent(route.path).absoluteString
 }
 
-private let lazyEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
-    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Closure({.NetworkResponse(200, {target.sampleData})}), method: target.method, parameters: target.parameters)
-}
-
 private let failureEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
     let error = NSError(domain: "com.moya.error", code: 0, userInfo: [NSLocalizedDescriptionKey: "Houston, we have a problem"])
-    return Endpoint<GitHub>(URL: url(target), sampleResponse: .NetworkError(error), method: target.method, parameters: target.parameters)
+    return Endpoint<GitHub>(URL: url(target), sampleResponseClosure: {.NetworkError(error)}, method: target.method, parameters: target.parameters)
 }
 
 private enum HTTPBin: MoyaTarget {