+++ Moya/Endpoint.swift
@@ -3,42 +3,36 @@ import Alamofire
 
 /// Used for stubbing responses.
 public enum EndpointSampleResponse {
-    // Swift won't let us put an enum inside a generic class like Endpoint.
 
-    case NetworkResponse(Int, () -> NSData)
+    /// The network returned a response, including status code and data.
+    case NetworkResponse(Int, NSData)
+
+    /// The network failed to send the request, or failed to retrieve a response (eg a timeout).
     case NetworkError(ErrorType?)
-    // TODO: We don't need this anymore, let's just evaluate all the sample responses.
-    case Closure(() -> EndpointSampleResponse)
-
-    func evaluate() -> EndpointSampleResponse {
-        switch self {
-        case NetworkResponse, NetworkError: return self
-        case Closure(let closure):
-            return closure().evaluate()
-        }
-    }
 }
 
 
 /// Class for reifying a target of the Target enum unto a concrete Endpoint.
 public class Endpoint<Target> {
+    public typealias SampleResponseClosure = () -> EndpointSampleResponse
+
     public let URL: String
     public let method: Moya.Method
-    public let sampleResponse: EndpointSampleResponse
+    public let sampleResponseClosure: SampleResponseClosure
     public let parameters: [String: AnyObject]?
     public let parameterEncoding: Moya.ParameterEncoding
     public let httpHeaderFields: [String: String]?
 
     /// Main initializer for Endpoint.
     public init(URL: String,
-        sampleResponse: EndpointSampleResponse,
+        sampleResponseClosure: SampleResponseClosure,
         method: Moya.Method = Moya.Method.GET,
         parameters: [String: AnyObject]? = nil,
         parameterEncoding: Moya.ParameterEncoding = .URL,
         httpHeaderFields: [String: String]? = nil) {
 
         self.URL = URL
-        self.sampleResponse = sampleResponse
+        self.sampleResponseClosure = sampleResponseClosure
         self.method = method
         self.parameters = parameters
         self.parameterEncoding = parameterEncoding
@@ -52,7 +46,7 @@ public class Endpoint<Target> {
             newParameters[key] = value
         }
 
-        return Endpoint(URL: URL, sampleResponse: sampleResponse, method: method, parameters: newParameters, parameterEncoding: parameterEncoding, httpHeaderFields: httpHeaderFields)
+        return Endpoint(URL: URL, sampleResponseClosure: sampleResponseClosure, method: method, parameters: newParameters, parameterEncoding: parameterEncoding, httpHeaderFields: httpHeaderFields)
     }
 
     /// Convenience method for creating a new Endpoint with the same properties as the receiver, but with added HTTP header fields.
@@ -62,13 +56,13 @@ public class Endpoint<Target> {
             newHTTPHeaderFields[key] = value
         }
 
-        return Endpoint(URL: URL, sampleResponse: sampleResponse, method: method, parameters: parameters, parameterEncoding: parameterEncoding, httpHeaderFields: newHTTPHeaderFields)
+        return Endpoint(URL: URL, sampleResponseClosure: sampleResponseClosure, method: method, parameters: parameters, parameterEncoding: parameterEncoding, httpHeaderFields: newHTTPHeaderFields)
     }
     
     /// Convenience method for creating a new Endpoint with the same properties as the receiver, but with another parameter encoding.
     public func endpointByAddingParameterEncoding(newParameterEncoding: Moya.ParameterEncoding) -> Endpoint<Target> {
         
-        return Endpoint(URL: URL, sampleResponse: sampleResponse, method: method, parameters: parameters, parameterEncoding: newParameterEncoding, httpHeaderFields: httpHeaderFields)
+        return Endpoint(URL: URL, sampleResponseClosure: sampleResponseClosure, method: method, parameters: parameters, parameterEncoding: newParameterEncoding, httpHeaderFields: httpHeaderFields)
     }
 }