+++ CHANGELOG.md
@@ -24,10 +24,12 @@
 - Removes inflight request tracking for ReactiveCocoa and RxSwift providers. **Breaking Change**
 - Adds support for ReactiveCocoa 4 by moving `ReactiveCocoaMoyaProvider` to use `SignalProducer` instead of `RACSignal`
 - Renamed `EndpointSampleResponse` cases: **Breaking Change**
-  - `Success` to `NetworkResponse`
+  - `Success` to `NetworkResponse`, now contains `NSData` instead of `() -> NSData`.
   - `Error` to `NetworkError`
   - Additionally, `NetworkError` no longer has a status code or data associated with it. This represents an error from the underlying iOS network stack, like an inability to connect. See [#200](https://github.com/Moya/Moya/issues/200) for more details.
-- ReactiveCocoa provider no longer replaces errors that contain status codes (an unlikely situation) with its own errors. It passes all errors directly through. 
+  - Also additionally, removed `Closure` case (see below).
+- Changed `Endpoint` to use a `sampleResponseClosure` instead of a `sampleResponse`, making all sample responses lazily executed. **Breaking Change**
+- ReactiveCocoa provider no longer replaces errors that contain status codes (an unlikely situation) with its own errors. It passes all errors directly through.
 
 # 2.4.1